v. 0.5.2  Изменение в оформлении кода
