+++ Course2FinalTask/Course2FinalTask/Controller/FeedCell/FeedCollectionViewCell.swift
@@ -12,22 +12,15 @@ import DataProvider
 class FeedCollectionViewCell: UICollectionViewCell, NibInit {
   
   @IBOutlet weak var avatarImageView: UIImageView!
-  
   @IBOutlet weak var userNameLable: UILabel!
-  
   @IBOutlet weak var dataLable: UILabel!
-  
   @IBOutlet weak var imageView: UIImageView!
-  
   @IBOutlet weak var likesLable: UILabel!
-  
-  
   @IBOutlet weak var descriptionLable: UILabel!
-  
   
   override func awakeFromNib() {
     super.awakeFromNib()
-//    imageView.contentMode = .scaleAspectFit
+    //    imageView.contentMode = .scaleAspectFit
   }
 }
