+++ Course2FinalTask/Course2FinalTask/Controller/ProfileViewController.swift
@@ -15,8 +15,4 @@ class ProfileViewController: UIViewController, NibInit {
 
 
     }
-
-
-   
-
 }
